LPS-186786 Exposing component to read batch engine units
