+++ modules/apps/batch-engine/batch-engine-service/src/main/java/com/liferay/batch/engine/internal/bundle/BatchEngineBundleTracker.java
@@ -14,23 +14,14 @@
 
 package com.liferay.batch.engine.internal.bundle;
 
-import com.liferay.batch.engine.unit.BatchEngineUnit;
 import com.liferay.batch.engine.unit.BatchEngineUnitProcessor;
+import com.liferay.batch.engine.unit.BatchEngineUnitReader;
 import com.liferay.petra.reflect.ReflectionUtil;
-import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.util.FileUtil;
-import com.liferay.portal.kernel.util.StringUtil;
 
 import java.io.File;
 import java.io.IOException;
 
-import java.net.URL;
-
-import java.util.Collection;
-import java.util.Dictionary;
-import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.Iterator;
 import java.util.Map;
 import java.util.Objects;
 
@@ -51,14 +42,6 @@ import org.osgi.util.tracker.BundleTrackerCustomizer;
 @Component(service = {})
 public class BatchEngineBundleTracker {
 
-	public boolean isBatchEngineTechnical(String zipEntryName) {
-		if (zipEntryName.endsWith(".batch-engine-data.json")) {
-			return true;
-		}
-
-		return false;
-	}
-
 	@Activate
 	protected void activate(
 		BundleContext bundleContext, Map<String, Object> properties) {
@@ -73,73 +56,6 @@ public class BatchEngineBundleTracker {
 	@Deactivate
 	protected void deactivate() {
 		_bundleTracker.close();
-	}
-
-	private String _getBatchEngineBundleEntryKey(URL url) {
-		String zipEntryName = url.getPath();
-
-		if (isBatchEngineTechnical(zipEntryName)) {
-			return zipEntryName;
-		}
-
-		if (!zipEntryName.contains(StringPool.SLASH)) {
-			return StringPool.BLANK;
-		}
-
-		return zipEntryName.substring(
-			0, zipEntryName.lastIndexOf(StringPool.SLASH));
-	}
-
-	private Collection<BatchEngineUnit> _getBatchEngineBundleUnitsCollection(
-		Bundle bundle, String batchPath) {
-
-		Map<String, URL> batchEngineURLs = new HashMap<>();
-		Map<String, BatchEngineUnit> batchEngineUnits = new HashMap<>();
-
-		Enumeration<URL> enumeration = bundle.findEntries(batchPath, "*", true);
-
-		while (enumeration.hasMoreElements()) {
-			URL url = enumeration.nextElement();
-
-			if (StringUtil.endsWith(url.getPath(), StringPool.SLASH)) {
-				continue;
-			}
-
-			String key = _getBatchEngineBundleEntryKey(url);
-
-			URL complementURL = batchEngineURLs.get(key);
-
-			if (complementURL == null) {
-				batchEngineURLs.put(key, url);
-
-				batchEngineUnits.put(
-					key, new AdvancedBatchEngineBundleUnitImpl(bundle, url));
-
-				continue;
-			}
-
-			batchEngineUnits.put(
-				key,
-				new ClassicBatchEngineBundleUnitImpl(
-					bundle, url, complementURL));
-
-			batchEngineURLs.remove(key);
-		}
-
-		return batchEngineUnits.values();
-	}
-
-	private Iterable<BatchEngineUnit> _getBatchEngineUnits(
-		Bundle bundle, String batchPath) {
-
-		return new Iterable<BatchEngineUnit>() {
-
-			@Override
-			public Iterator<BatchEngineUnit> iterator() {
-				return new BatchEngineUnitIterator(bundle, batchPath);
-			}
-
-		};
 	}
 
 	private boolean _isAlreadyProcessed(Bundle bundle) {
@@ -169,36 +85,11 @@ public class BatchEngineBundleTracker {
 		return false;
 	}
 
-	private void _processBatchEngineBundle(Bundle bundle) {
-		Dictionary<String, String> headers = bundle.getHeaders(
-			StringPool.BLANK);
-
-		String batchPath = headers.get("Liferay-Client-Extension-Batch");
-
-		if (batchPath != null) {
-			if (_isAlreadyProcessed(bundle)) {
-				return;
-			}
-
-			if (batchPath.isEmpty()) {
-				batchPath = StringPool.PERIOD;
-			}
-
-			if (StringUtil.startsWith(batchPath, StringPool.SLASH)) {
-				batchPath = batchPath.substring(1);
-			}
-
-			if (!StringUtil.endsWith(batchPath, StringPool.SLASH)) {
-				batchPath = batchPath.concat(StringPool.SLASH);
-			}
-
-			_batchEngineUnitProcessor.processBatchEngineUnits(
-				_getBatchEngineUnits(bundle, batchPath));
-		}
-	}
-
 	@Reference(policyOption = ReferencePolicyOption.GREEDY)
 	private BatchEngineUnitProcessor _batchEngineUnitProcessor;
+
+	@Reference(policyOption = ReferencePolicyOption.GREEDY)
+	private BatchEngineUnitReader _batchEngineUnitReader;
 
 	private BundleTracker<Void> _bundleTracker;
 
@@ -207,7 +98,10 @@ public class BatchEngineBundleTracker {
 
 		@Override
 		public Void addingBundle(Bundle bundle, BundleEvent bundleEvent) {
-			_processBatchEngineBundle(bundle);
+			if (!_isAlreadyProcessed(bundle)) {
+				_batchEngineUnitProcessor.processBatchEngineUnits(
+					_batchEngineUnitReader.getBatchEngineUnits(bundle));
+			}
 
 			return null;
 		}
@@ -221,29 +115,6 @@ public class BatchEngineBundleTracker {
 		public void removedBundle(
 			Bundle bundle, BundleEvent bundleEvent, Void unused) {
 		}
-
-	}
-
-	private class BatchEngineUnitIterator implements Iterator<BatchEngineUnit> {
-
-		public BatchEngineUnitIterator(Bundle bundle, String batchPath) {
-			Collection<BatchEngineUnit> batchEngineZipUnits =
-				_getBatchEngineBundleUnitsCollection(bundle, batchPath);
-
-			_iterator = batchEngineZipUnits.iterator();
-		}
-
-		@Override
-		public boolean hasNext() {
-			return _iterator.hasNext();
-		}
-
-		@Override
-		public BatchEngineUnit next() {
-			return _iterator.next();
-		}
-
-		private final Iterator<BatchEngineUnit> _iterator;
 
 	}
